pyproject.toml
[project]
authors = [
    {name = "Oleg Dixon", email = "olegdixon5973@gmail.com"},
]
name = "hexlet-code"
version = "0.1.0"
description = "Difference Calculator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-cov>=6.1.1",
    "PyYAML>=6.0.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.1",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.main:main"
difference = "gendiff.main:main"
